move kinematic model to own module
