+++ src/lib.rs
@@ -1,6 +1,7 @@
 mod euler; // Add this line to import the new module
-
+mod kinematic_model; // Add this line to import the new module
 use crate::euler::EulerConvention;
+use crate::kinematic_model::KinematicModel;
 
 use nalgebra::{Isometry3, Rotation3, Translation3, Vector3};
 use pyo3::exceptions::PyValueError;
@@ -13,151 +14,6 @@ use polars::series::Series;
 use pyo3_polars::PyDataFrame;
 use rs_opw_kinematics::kinematic_traits::{Kinematics, Pose};
 use rs_opw_kinematics::kinematics_impl::OPWKinematics;
-use rs_opw_kinematics::parameters::opw_kinematics::Parameters;
-
-#[pyclass(frozen)] // Declare the class as frozen to provide immutability.
-#[derive(Clone)]
-struct KinematicModel {
-    a1: f64,
-    a2: f64,
-    b: f64,
-    c1: f64,
-    c2: f64,
-    c3: f64,
-    c4: f64,
-    offsets: [f64; 6],
-    flip_axes: [bool; 6], // Renamed and changed to boolean array
-    has_parallelogram: bool,
-}
-
-impl KinematicModel {
-    fn to_opw_kinematics(&self, degrees: bool) -> OPWKinematics {
-        let sign_corrections = self.flip_axes.map(|x| if x { -1 } else { 1 });
-        OPWKinematics::new(Parameters {
-            a1: self.a1,
-            a2: self.a2,
-            b: self.b,
-            c1: self.c1,
-            c2: self.c2,
-            c3: self.c3,
-            c4: self.c4,
-            offsets: if degrees {
-                self.offsets
-                    .iter()
-                    .map(|&x| x.to_radians())
-                    .collect::<Vec<_>>()
-                    .try_into()
-                    .unwrap()
-            } else {
-                self.offsets
-            },
-            sign_corrections,
-            dof: 6,
-        })
-    }
-}
-
-#[pymethods]
-impl KinematicModel {
-    #[new]
-    #[pyo3(signature = (
-        a1 = 0.0,
-        a2 = 0.0,
-        b = 0.0,
-        c1 = 0.0,
-        c2 = 0.0,
-        c3 = 0.0,
-        c4 = 0.0,
-        offsets = (0.0, 0.0, 0.0, 0.0, 0.0, 0.0),
-        flip_axes = (false, false, false, false, false, false),
-        has_parallelogram = false
-    ))]
-    #[allow(clippy::too_many_arguments)]
-    fn new(
-        a1: f64,
-        a2: f64,
-        b: f64,
-        c1: f64,
-        c2: f64,
-        c3: f64,
-        c4: f64,
-        offsets: (f64, f64, f64, f64, f64, f64),
-        flip_axes: (bool, bool, bool, bool, bool, bool),
-        has_parallelogram: bool,
-    ) -> PyResult<Self> {
-        Ok(KinematicModel {
-            a1,
-            a2,
-            b,
-            c1,
-            c2,
-            c3,
-            c4,
-            offsets: offsets.into(),
-            flip_axes: flip_axes.into(),
-            has_parallelogram,
-        })
-    }
-
-    // Getter methods to provide access to attributes since the class is frozen.
-    #[getter]
-    fn a1(&self) -> f64 {
-        self.a1
-    }
-
-    #[getter]
-    fn a2(&self) -> f64 {
-        self.a2
-    }
-
-    #[getter]
-    fn b(&self) -> f64 {
-        self.b
-    }
-
-    #[getter]
-    fn c1(&self) -> f64 {
-        self.c1
-    }
-
-    #[getter]
-    fn c2(&self) -> f64 {
-        self.c2
-    }
-
-    #[getter]
-    fn c3(&self) -> f64 {
-        self.c3
-    }
-
-    #[getter]
-    fn c4(&self) -> f64 {
-        self.c4
-    }
-
-    #[getter]
-    fn offsets(&self) -> Vec<f64> {
-        self.offsets.to_vec() // Convert the array to a Vec for easier handling in Python.
-    }
-
-    #[getter]
-    fn flip_axes(&self) -> Vec<bool> {
-        self.flip_axes.to_vec() // Convert the array to a Vec for easier handling in Python.
-    }
-
-    #[getter]
-    fn has_parallelogram(&self) -> bool {
-        self.has_parallelogram
-    }
-
-    fn __repr__(&self) -> String {
-        format!(
-            "KinematicModel(\n    a1={},\n    a2={},\n    b={},\n    c1={},\n    c2={},\n    c3={},\n    c4={},\n    offsets={:?},\n    flip_axes={:?},\n    has_parallelogram={}\n)",
-            self.a1, self.a2, self.b, self.c1, self.c2, self.c3, self.c4,
-            self.offsets, self.flip_axes, self.has_parallelogram
-        )
-    }
-}
 
 #[pyclass]
 struct Robot {
